added abstract CommandActivity
